Better way to handle text editing acitons
