+++ CoreEditor/src/bridge/web/selection.ts
@@ -1,6 +1,6 @@
 import { WebModule } from '../webModule';
 import { WebRect } from '../../@types/WebRect';
-import { selectedMainText, selectAll, scrollToSelection, getRect, gotoLine, refreshEditFocus } from '../../modules/selection';
+import { selectedMainText, scrollToSelection, getRect, gotoLine, refreshEditFocus } from '../../modules/selection';
 
 /**
  * @shouldExport true
@@ -10,7 +10,6 @@ import { selectedMainText, selectAll, scrollToSelection, getRect, gotoLine, refr
 export interface WebModuleSelection extends WebModule {
   getText(): string;
   getRect({ pos }: { pos: CodeGen_Int }): WebRect | undefined;
-  selectAll(): void;
   scrollToSelection(): void;
   gotoLine({ lineNumber }: { lineNumber: CodeGen_Int }): void;
   refreshEditFocus(): void;
@@ -23,10 +22,6 @@ export class WebModuleSelectionImpl implements WebModuleSelection {
 
   getRect({ pos }: { pos: CodeGen_Int }): WebRect | undefined {
     return getRect(pos);
-  }
-
-  selectAll(): void {
-    selectAll();
   }
 
   scrollToSelection(): void {
